utils/requirement.txt
groq==0.4.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
python-magic==0.4.27
pathlib==1.0.1
pytest==7.4.3
black==23.11.0
flake8==6.1.0